extract function to a utility
