+++ src/javascript/node-project.ts
@@ -6,7 +6,7 @@ import type { JestOptions } from "./jest";
 import { Jest } from "./jest";
 import type { LicenseCheckerOptions } from "./license-checker";
 import { LicenseChecker } from "./license-checker";
-import type { CodeArtifactOptions, NodePackageOptions } from "./node-package";
+import type { NodePackageOptions } from "./node-package";
 import {
   CodeArtifactAuthProvider as NodePackageCodeArtifactAuthProvider,
   NodePackage,
@@ -29,12 +29,12 @@ import type { BiomeOptions } from "./biome/biome";
 import { Biome } from "./biome/biome";
 import {
   execCommand,
-  executeCommandPriorInstallation,
+  getScopedPackageWorkflowSteps,
   isYarnBerry,
   isYarnClassic,
 } from "./util";
 import { DEFAULT_GITHUB_ACTIONS_USER } from "../github/constants";
-import { ensureNotHiddenPath, secretToString } from "../github/private/util";
+import { ensureNotHiddenPath } from "../github/private/util";
 import type {
   JobPermissions,
   JobStep,
@@ -1143,87 +1143,6 @@ export class NodeProject extends GitHubProject {
   }
 
   /**
-   * Get steps for scoped package access
-   *
-   * @param codeArtifactOptions Details of logging in to AWS
-   * @returns array of job steps required for each private scoped packages
-   */
-  public getScopedPackageSteps(
-    codeArtifactOptions: CodeArtifactOptions | undefined,
-  ): JobStep[] {
-    const parsedCodeArtifactOptions = {
-      accessKeyIdSecret:
-        codeArtifactOptions?.accessKeyIdSecret ?? "AWS_ACCESS_KEY_ID",
-      secretAccessKeySecret:
-        codeArtifactOptions?.secretAccessKeySecret ?? "AWS_SECRET_ACCESS_KEY",
-      roleToAssume: codeArtifactOptions?.roleToAssume,
-      authProvider: codeArtifactOptions?.authProvider,
-    };
-
-    const executeProjenCommand = `${executeCommandPriorInstallation(this.packageManager)} projen`;
-
-    if (
-      parsedCodeArtifactOptions.authProvider ===
-      NodePackageCodeArtifactAuthProvider.GITHUB_OIDC
-    ) {
-      return [
-        {
-          name: "Configure AWS Credentials",
-          uses: "aws-actions/configure-aws-credentials@v6",
-          with: {
-            "aws-region": "us-east-2",
-            "role-to-assume": parsedCodeArtifactOptions.roleToAssume,
-            "role-duration-seconds": 900,
-          },
-        },
-        {
-          name: "AWS CodeArtifact Login",
-          run: `${executeProjenCommand} ca:login`,
-        },
-      ];
-    }
-
-    if (parsedCodeArtifactOptions.roleToAssume) {
-      return [
-        {
-          name: "Configure AWS Credentials",
-          uses: "aws-actions/configure-aws-credentials@v6",
-          with: {
-            "aws-access-key-id": secretToString(
-              parsedCodeArtifactOptions.accessKeyIdSecret,
-            ),
-            "aws-secret-access-key": secretToString(
-              parsedCodeArtifactOptions.secretAccessKeySecret,
-            ),
-            "aws-region": "us-east-2",
-            "role-to-assume": parsedCodeArtifactOptions.roleToAssume,
-            "role-duration-seconds": 900,
-          },
-        },
-        {
-          name: "AWS CodeArtifact Login",
-          run: `${executeProjenCommand} ca:login`,
-        },
-      ];
-    }
-
-    return [
-      {
-        name: "AWS CodeArtifact Login",
-        run: `${executeProjenCommand} ca:login`,
-        env: {
-          AWS_ACCESS_KEY_ID: secretToString(
-            parsedCodeArtifactOptions.accessKeyIdSecret,
-          ),
-          AWS_SECRET_ACCESS_KEY: secretToString(
-            parsedCodeArtifactOptions.secretAccessKeySecret,
-          ),
-        },
-      },
-    ];
-  }
-
-  /**
    * Returns the set of workflow steps which should be executed to bootstrap a
    * workflow.
    *
@@ -1286,7 +1205,10 @@ export class NodeProject extends GitHubProject {
 
     if (this.package.scopedPackagesOptions) {
       install.push(
-        ...this.getScopedPackageSteps(this.package.codeArtifactOptions),
+        ...getScopedPackageWorkflowSteps(
+          this.package.packageManager,
+          this.package.codeArtifactOptions,
+        ),
       );
     }
